json_set_val method is completed
added all parameter for freeing top elements of the element to json_free
number type of a json value changed to double from long

+++ include/discord.h
@@ -37,7 +37,7 @@
 	struct Embed {
 		char *title;
 		char *description;
-		unsigned int color;
+		double color;
 		char *image_url;
 		char *thumbnail_url;
 		struct EmbedAuthor author;
